requirements.txt
Pillow
pyautogui
